pyproject.toml
[project]
name = "pdf-to-excel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.51.0",
    "google-generativeai>=0.8.5",
    "ocrmypdf>=16.12.0",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.6",
    "pypdf2>=3.0.1",
    "tesseract>=0.1.3",
    "ttkbootstrap>=1.19.0",
    "xlsxwriter>=3.2.9",
]
